ファイルを見てシンボリックリンクを貼るようにした

+++ scripts/link.sh
@@ -4,7 +4,12 @@ DOT_FILES=( .zshrc .zshenv .gitconfig .vimrc .vim .ssh bin .tmux.conf)
 # シンボリックリンクの作成
 for file in ${DOT_FILES[@]}
 do
-  ln -s -i $HOME/dotfiles/$file $HOME/$file
-  echo "ln -s -i $HOME/dotfiles/$file $HOME/$file"
+  if [ -e $HOME/$file ]; then
+    echo "already exists $HOME/$file"
+  else
+      # 存在しない場合
+    ln -s -i $HOME/dotfiles/$file $HOME/$file
+    echo "create symbolic link $HOME/dotfiles/$file $HOME/$file"
+  fi
 done
 echo "linked dotfiles complete!"
